Commits base of new syntax checker, which is still in its infant stage. This checker is EXTREMELY LENIENT, to the point where it will miss some glaring errors. Please confirm it does NOT error on valid code; that is the point of this test. Fixes nested switch. Fixes ten different bugs with the compiler from TGMG's list. Fixes max() and other varargs functions temporarily. Fixes room=0. Fixes non-GNU (ISO) ambiguity.
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@822 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Universal_System/highscore_functions.cpp
@@ -171,7 +171,7 @@ int string_width(variant str);
 
 void draw_highscore(int x1, int y1, int x2, int y2) {
 
-    for (int i=0; i<enigma::highscore_list.size(); i++) {
+    for (size_t i=0; i<enigma::highscore_list.size(); i++) {
         draw_text(x1, (i*((y2-y1)/10))+y1+10, enigma::highscore_list[i].player_name.c_str());
         draw_text(x2-string_width(toString((var)enigma::highscore_list[i].player_score)), (i*((y2-y1)/10))+y1+10 , toString((var)enigma::highscore_list[i].player_score));
     }
